user owned lists, user initals, logout tweak implemented "fixing commit names"

+++ server/controllers/top5list-controller.js
@@ -50,6 +50,11 @@ updateTop5List = async (req, res) => {
                 message: 'Top 5 List not found!',
             })
         }
+        if(top5List.ownerEmail !== req.params.email){
+            return res.status(401).json({ err, 
+                message: 'User does not have ownership of the Top 5 List!'
+            }); 
+        }
 
         top5List.name = body.name
         top5List.items = body.items
@@ -81,6 +86,11 @@ deleteTop5List = async (req, res) => {
                 message: 'Top 5 List not found!',
             })
         }
+        if(top5List.ownerEmail !== req.params.email){
+            return res.status(401).json({ err, 
+                message: 'User does not have ownership of the Top 5 List!'
+            }); 
+        }
         Top5List.findOneAndDelete({ _id: req.params.id }, () => {
             return res.status(200).json({ success: true, data: top5List })
         }).catch(err => console.log(err))
@@ -91,6 +101,10 @@ getTop5ListById = async (req, res) => {
     await Top5List.findById({ _id: req.params.id }, (err, list) => {
         if (err) {
             return res.status(400).json({ success: false, error: err });
+        }
+        if(list.ownerEmail !== req.params.email){
+            return res.status(401).json({ err, message: 
+                'User does not have ownership of the Top 5 List!'}); 
         }
         return res.status(200).json({ success: true, top5List: list })
     }).catch(err => console.log(err))
@@ -105,7 +119,8 @@ getTop5Lists = async (req, res) => {
                 .status(404)
                 .json({ success: false, error: `Top 5 Lists not found` })
         }
-        return res.status(200).json({ success: true, data: top5Lists })
+        let ownedTop5Lists = top5Lists.filter(list => list.ownerEmail === req.params.email);
+        return res.status(200).json({ success: true, data: ownedTop5Lists })
     }).catch(err => console.log(err))
 }
 getTop5ListPairs = async (req, res) => {
@@ -124,6 +139,9 @@ getTop5ListPairs = async (req, res) => {
             let pairs = [];
             for (let key in top5Lists) {
                 let list = top5Lists[key];
+                if(list.ownerEmail !== req.params.email){
+                    continue; 
+                }
                 let pair = {
                     _id: list._id,
                     name: list.name
